Change the output methods
Change the output methods to have the byte output as the default, the matlab output seperate from the byte output (because it eats its arrays differently). PythonOutput is no longer for python, but now functions more as a text output for "debugging" purposes.

Change the project file to make the Debug version a multithreaded DLL instead of MT, as the latter option doesn't work.