Calculation of average fall velocity and time until steady concentration.
Use --avgfallvelocity and --timesteady to enable the calculation. Results are printed to stdout.

Average fall velocity is calculated by adding the fall velocity of each particle to a total when it leaves the box at the bottom, and in the end divide the total sum of fall velocities by the amount of particles that left the box at the bottom.

Steady concentration is calculated by doing a 2-pass, with the first pass being the "regular" one which may output to a file, but the second pass calculated the concentration every --outputinterval and gets the error between the current concentration and the final concentration of the last pass using a Frobenius norm of the difference of the two.
--error-k sets the threshold (default is 0.01).

- getConcentration() in Output no longer receives the ScalarField it should write to, but instead makes and returns one itself.
- reset() in Emitter no longer returns anything ( life time calculation is done in the checkParticles() section in Main.cpp ).
- outsideBox() in Vortex can return multiple values. This is to separate the case where the particle left at the bottom from the other cases.
- add functions in Particle that make reading of the startPos possible.